优化37

+++ app/src/main/java/com/ski/box/bean/BetStatus.java
@@ -2,8 +2,8 @@ package com.ski.box.bean;
 
 public class BetStatus {
     private int status; // 0 单式  1 组合
-    private int totalAmount;
-    private float totalAmount_s;
+    private long totalAmount;
+    private double totalAmount_s;
     private int zhuShu = 0;
 
     public BetStatus() {
@@ -23,19 +23,19 @@ public class BetStatus {
         this.status = status;
     }
 
-    public int getTotalAmount() {
+    public long getTotalAmount() {
         return totalAmount;
     }
 
-    public void setTotalAmount(int totalAmount) {
+    public void setTotalAmount(long totalAmount) {
         this.totalAmount = totalAmount;
     }
 
-    public float getTotalAmount_s() {
+    public double getTotalAmount_s() {
         return totalAmount_s;
     }
 
-    public void setTotalAmount_s(float totalAmount_s) {
+    public void setTotalAmount_s(double totalAmount_s) {
         this.totalAmount_s = totalAmount_s;
     }
